创建版本 1.0.2.

+++ NumberSwitchString/NumberSwitchString.h
@@ -11,7 +11,7 @@
 NS_ASSUME_NONNULL_BEGIN
 
 @interface NumberSwitchString : NSObject
-+(NSString *)numberSwitchString:(NSInteger )number;
++(NSString *)numberSwitchString:(long )number;
 +(NSString *)objSwitchString:(NSObject *)obj;
 + (BOOL)isNumber:(NSString *)strValue;
 +(NSInteger)stringSwitchNumber:(NSString *)numberString;
