+++ NumberSwitchString/NumberSwitchString.m
@@ -7,10 +7,8 @@
 //
 
 #import "NumberSwitchString.h"
-#import <UIKit/UIKit.h>
-#import <Foundation/Foundation.h>
 @implementation NumberSwitchString
-+(NSString *)numberSwitchString:(NSInteger )number
++(NSString *)numberSwitchString:(long )number
 {
     return [NSString stringWithFormat:@"%ld",number];
 }
